BCG

Description

Library for handling bi-coloured graphs

Data Types

data BCG

Bi-Colored Graph

Constructors

BCG

Fields

[BCE]

List of BCEs

IntSet

Set of Vertices

Int

Number of different colours

data BCE

Bi-Coloured Edge

Constructors

BCE

Fields

fromV :: Int

From Vertex

fromC :: Int

From Colour

toV :: Int

To Vertex

toC :: Int

To Colour

w :: Complex Double

Weight

type IVC = Set (Int, Int)

Inherited Vertex Coloring

type PM = BCG

Perfect Matching

Constructors

empty :: BCG

A BCG with no edges or vertices and the number of different colors is 0

Functions

dist :: BCG -> Double

Return a value between 0 and 1 representing a BCG's distance to the monochromatic \[dist(G) = \begin{cases} 0 & G \text{ has no perfect matchings} \\ \frac{\left|\displaystyle\sum_{c_i \in G,\ c_i\ \text{is monochromatic}} w(c_i)\right|^2}{d\cdot\left(\displaystyle\sum_{c_i \in G} |w(c_i)|^2\right)} & \text{Otherwise} \end{cases}\] Where \(c_i\) is an IVC on \(G\), \(d\) is the number of different colors on \(G\) and \(w(c_i)\) is the colouring weight of \(c_i\).

enumeratePM :: BCG -> [BCG]

Get a list of perfect matchings on a BCG The perfect matchings are represented by the BCG data type

getIncidentEdges :: BCG -> Int -> [BCE]

Get a list of BCEs connected to a vertex

isIncident :: Int -> BCE -> Bool

Determines if a vertex is connected to an BCE

removeEdgeAndIncidentVertices :: BCG -> BCE -> BCG

Removes a BCE from a BCG and the vertices it was incident to

removeVertex :: BCG -> Int -> BCG

Removes a vertex and the connected BCEs from a BCG

removeVertices :: BCG -> [Int] -> BCG

Removes a list of vertices and their connected BCEs from a BCG

isMonochromatic :: PM -> Bool

Checks if all edges on a PM are monochromatic and have the same color

coloringWeight :: [PM] -> Complex Double

Takes a list of PMs with the same IVC and returns the IVC's weight

pmsIvc :: PM -> IVC

Get the IVC of a PM

groupByIvc :: [PM] -> [[PM]]

Groups a list of PM's by IVC

isAdjacent :: BCG -> Int -> Int -> Bool

Check if 2 vertices are adjacent

getAdjacent :: BCG -> Int -> IntSet

Get a set of adjacent nodes

getConnected :: BCG -> Int -> IntSet

get list of vertices connected to a vertex on a BCG

isConnectedGraph :: BCG -> Bool

Determine if a given BCG is a fully connected graph

numberDifferentColors :: BCG -> Int

Counts the number of different colors on a BCG